Created endpoints, validators and redis

+++ backend-java/src/main/java/kafka/system/br/AutoFrota/model/Maintenance.java
@@ -28,7 +28,7 @@ public class Maintenance {
     private boolean scheduled;
 
     @Column(name = "total_value", nullable = false)
-    private float totalValue;
+    private Double totalValue;
 
     @ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     @JoinColumn(name = "vehicle_id")
@@ -40,7 +40,7 @@ public class Maintenance {
     public Maintenance() {
     }
 
-    public Maintenance(Long id, Date date, boolean done, String observation, boolean scheduled, float totalValue,Vehicle vehicle) {
+    public Maintenance(Long id, Date date, boolean done, String observation, boolean scheduled, Double totalValue,Vehicle vehicle) {
         this.id = id;
         this.date = date;
         this.done = done;
@@ -50,7 +50,7 @@ public class Maintenance {
         this.vehicle = vehicle;
     }
 
-    public Maintenance(Date date, boolean done, String observation, boolean scheduled, float totalValue, Vehicle vehicle) {
+    public Maintenance(Date date, boolean done, String observation, boolean scheduled, Double totalValue, Vehicle vehicle) {
         this.date = date;
         this.done = done;
         this.observation = observation;
@@ -99,11 +99,11 @@ public class Maintenance {
         this.scheduled = scheduled;
     }
 
-    public float getTotalValue() {
+    public Double getTotalValue() {
         return totalValue;
     }
 
-    public void setTotalValue(float totalValue) {
+    public void setTotalValue(Double totalValue) {
         this.totalValue = totalValue;
     }
